Moved the add_domain function
