+++ mmdet/utils/__init__.py
@@ -1,7 +1,6 @@
 from .collect_env import collect_env
 from .flops_counter import get_model_complexity_info
 from .logger import get_root_logger
-from .misc import ExtendedDictAction, add_domain, DOMAIN_CUSTOM_OPS_NAME
+from .misc import ExtendedDictAction
 
-__all__ = ['get_model_complexity_info', 'get_root_logger', 'collect_env', 'ExtendedDictAction',
-           'add_domain', 'DOMAIN_CUSTOM_OPS_NAME']
+__all__ = ['get_model_complexity_info', 'get_root_logger', 'collect_env', 'ExtendedDictAction']
