+++ mmdet/utils/deployment/operations_domain.py
@@ -1,0 +1,4 @@
+DOMAIN_CUSTOM_OPS_NAME = 'org.openvinotoolkit'
+
+def add_domain(name_operator: str) -> str:
+    return DOMAIN_CUSTOM_OPS_NAME + '::' + name_operator
