+++ mmdet/utils/misc.py
@@ -2,12 +2,6 @@ import ast
 from mmcv import DictAction
 
 
-DOMAIN_CUSTOM_OPS_NAME = 'org.openvinotoolkit'
-
-def add_domain(name_operator: str) -> str:
-    return DOMAIN_CUSTOM_OPS_NAME + '::' + name_operator
-
-
 class ExtendedDictAction(DictAction):
     """
     argparse action to split an argument into KEY=VALUE form
